+++ src/org/genesys/ml/YFeaturizer.java
@@ -3,5 +3,5 @@ package org.genesys.ml;
 import java.util.List;
 
 public interface YFeaturizer {
-	public List<Double> getFeatures(String function);
+	public List<Integer> getFeatures(String function);
 }
